Fix enums in json-format output; emit emitter_category as a string.
Disable clang-format of the serialization maps to make them readable.
Put the serialization in the right namespace, so it actually works.
Add an INVALID enum value to all enums for completeness if we ever want to do deserialization.
Emit emitter_category as a string like dump1090 uses - set letter + category (A0, B3, etc)